requirements.txt
accelerate==0.25.0
aiohttp==3.9.1
aiosignal==1.3.1
anthropic==0.7.7
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
aspy.yaml==1.3.0
astor==0.8.1
astroid==2.5.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blis==0.7.11
cached-property==1.5.2
catalogue==2.0.10
certifi==2020.12.5
cffi==1.16.0
cfgv==3.2.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
comm==0.2.0
confection==0.1.4
cymem==2.0.8
Cython==3.0.6
dataclasses==0.6
datasets==2.15.0
debugpy==1.8.0
decorator==4.4.2
deepspeed==0.12.4
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.1
distro==1.8.0
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
faiss-gpu==1.7.2
fastapi==0.104.1
fastjsonschema==2.19.0
filelock==3.0.12
flatbuffers==23.5.26
fqdn==1.5.1
frozenlist==1.4.0
fschat==0.2.28
fsspec==2023.10.0
ftfy==6.1.3
gast==0.4.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
identify==2.1.0
idna==2.10
importlib-metadata==7.0.0
importlib-resources==6.1.1
iniconfig==1.1.1
ipykernel==6.27.1
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.7.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
langcodes==3.3.0
lazy-object-proxy==1.5.2
lightgbm==4.1.0
markdown-it-py==3.0.0
markdown2==2.4.11
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.6.1
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nh3==0.2.15
ninja==1.11.1.1
nltk==3.8.1
nmslib==2.1.1
nodeenv==1.5.0
notebook==6.4.12
notebook_shim==0.2.3
numpy==1.24.4
onnxruntime==1.16.3
openai==0.27.8
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec @ file:///root/paddlejob/workspace/yihang/package/pathspec-0.12.1-py3-none-any.whl#sha256=a0d503e138a4c123b27490a4f7beda6a01c6f288df0e4a8b79c7eb0dc7b4cc08
peft==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.1
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
pluggy==0.13.1
pre-commit==1.10.4
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==3.15.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pybind11==2.6.1
pycparser==2.21
pydantic==1.10.13
Pygments==2.17.2
PyGObject==3.26.1
pyjnius==1.6.1
pylint==2.7.2
pynvml==11.5.0
pyparsing==2.4.7
pyserini==0.22.1
pytest==6.2.2
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==5.4.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rouge==1.0.1
rpds-py==0.13.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.2
sentencepiece==0.1.99
setuptools-scm==8.0.4
shortuuid==1.0.11
simplegeneric==0.8.1
six==1.15.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
ssh-import-id==5.7
stack-data==0.6.3
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
terminado==0.18.0
tevatron==0.1.0
thinc==8.2.1
threadpoolctl==3.2.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
torch @ file:///root/paddlejob/workspace/yihang/torch-2.1.0%2Bcu118-cp38-cp38-linux_x86_64.whl#sha256=02cd2c312501ebd9faf65bedb48ffbff77312ffef04cf7125ed4caa1738fd8df
tornado==6.4
tqdm==4.66.1
traitlets==5.6.0
transformers==4.35.2
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2023.3
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.3
uvicorn==0.24.0.post1
virtualenv==20.4.2
wasabi==1.1.2
wavedrom==2.0.3.post3
wcwidth==0.2.12
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
wrapt==1.12.1
xformers @ file:///root/paddlejob/workspace/yihang/package/xformers-0.0.22.post7%2Bcu118-cp38-cp38-manylinux2014_x86_64.whl#sha256=27effefc6ce76362d88897ffdb7b1364290a9773d37da92fbe2e06470067cfcf
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
